refactor(llment): add prompts module
